Created + Tested Templated Linked Lists
I created a basic template for making Linked Lists which include functions like creating a node, adding a node, and deleting a node. It is fully functional and fully tested. The linked lists work as intended. The pointers also work as intended and as the goal of this bit of work was practicing linked lists and pointers this was a good way of doing both. I am more able to work with Linked Lists and Pointers.
To Do:
1. Practicing binary trees is next with a templated example like this. I tried jumping into the deep end with an example but it did not work as intended. This shows a simpler example that worked.
2. Redo the git log as a Linked List.